// server/types/enumType.ts
export enum SymbolEnum {
  BTCUSDT = "btcusdt",
  BNBBTC = "bnbbtc",
  BNBBUSD = "bnbbusd",
  ETHBUSD = "ethbusd",
  XBTUSD = "XBTUSD",
}

export enum TimeEnum {
  oneMin = "1m",
  threeMin = "3m",
  fiveMin = "5m",
  fifteenMin = "15m",
  thirtyMin = "30m",
  oneHour = "1h",
  oneDay = "1d",
}

export enum ExchangeEnum {
  BINANCE = "BINANCE",
  BITMEX = "BITMEX",
}

export enum TypeEnum {
  KLINE = "KLINE",
  ORDER = "ORDER",
}